requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyyaml>=6.0

# API clients
requests>=2.31.0

# Database
pymongo>=4.6.0
motor>=3.3.0  # Async MongoDB driver

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Web Application
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Dependencies for pandas
pytz
python-dateutil
